% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{meanKRS_cpp}
\alias{meanKRS_cpp}
\title{Estimate conditional expectation using a Gaussian Kernel with fixed variance.}
\usage{
meanKRS_cpp(y, x, x0, lambda)
}
\arguments{
\item{y}{numeric vector}

\item{x}{numeric vector}

\item{x0}{numeric vector}

\item{lambda}{double}
}
\value{
conditional expectation
}
\description{
Estimate conditional expectation using a Gaussian Kernel with fixed variance.
}
