% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{varKRS_cpp}
\alias{varKRS_cpp}
\title{Estimate conditional expectation using a Gaussian Kernel with changing bandwidth.}
\usage{
varKRS_cpp(y, x, x0, lambda)
}
\arguments{
\item{y}{numeric vector}

\item{x}{numeric vector}

\item{x0}{numeric vector}

\item{lambda}{double}
}
\value{
conditional expectation
}
\description{
Estimate conditional expectation using a Gaussian Kernel with changing bandwidth.
}
